Toggle no identical checkbox added
In this update, we added a checkbox for toggling whether or not identical strings are added to the final output. We accomplished this by making a slightly modified loop that checks if there is already an existing instance of each string. This modified loop is only used if the user checks a checkbox,